feat: Add ability for Reactotron to communicate with Flipper
BREAKING CHANGES: The way communication is handled has changed with this release to add the ability to communicate with both Reactotron standalone and Reactotron in Flipper. The only API change is if you intend to use Reactotron in Flipper. If you do here are the steps that you need to take:

1. Install at least this version of `reactotron-react-native`
2. Install `react-native-flipper`
3. Have Flipper configured in your app (`react-native` template does this starting in 0.62.0)
4. `pod install` within the `ios` folder of your project
5. Add the following to your standard reactotron config:

```
import ReactotronFlipper from 'reactotron-react-native/dist/flipper';

{...}

Reactotron.configure({
  createSocket: path => new ReactotronFlipper(path), // You can have the other options but this one at a minimum is required.
});
```