feat(reactotron-app): Timeline deep-search (#1417)
This PR adapts @maxkomarychev's pull request #1181. 

The codebase has changed significantly since the PR was opened so this
is a reimplementation of that same idea: if it's a string, search it. if
it's not, try to stringify the object and search that string for
whatever the user has typed into the timeline search box.

Can I get a couple reviewers to try this out with like an hours worth of
reactotron timeline logs to see if this adversely affects performance in
a real-world environment with lots of logs to search?

Before: 


![before](https://github.com/infinitered/reactotron/assets/139261/3350197c-4417-427b-ae9a-78d2fd322d6c)

After:


![after](https://github.com/infinitered/reactotron/assets/139261/9cc4a2e2-6d47-4891-b97e-0992cca6e7d6)