Upgrade Example App to Expo SDK 50 & RN 0.73.6 (#1460)
## Please verify the following:

- [x] `yarn build-and-test:local` passes
- [ ] ~I have added tests for any new features, if relevant~
- [ ] ~`README.md` (or relevant documentation) has been updated with
your changes~

## Describe your PR
This PR upgrades the example application to Expo SDK 50 and React Native
0.73.6.


### Additional Notes
- `console.tron.image` relies on placekitten.com which has been
struggling as of late. Vote to switch that out with something that we
control.
- When developing on Reactotron the `reactotron-react-native` plugin has
to have its `devDependency` set to the same version as the example
application. This was not intuitive and discovered through
`XHRInterceptor` ceasing to function as expected.